+++ src/components/index.js
@@ -5,4 +5,5 @@ export Dropdown from './Dropdown';
 export OverflowMenu from './OverflowMenu';
 export Icon from './Icon';
 export Table from './Table';
+export LoadingSpinner from './LoadingSpinner';
 export Searchbar from './Searchbar';