+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 ### Added
 - Searchbar component
+- Loading spinner component
 - Table component
 - Overflow menu component
 - ToggleMenu component